start screen stuff.

+++ Avg4k-refactor/avg4k-new/avg4k-new/GameObject.h
@@ -41,7 +41,7 @@ namespace AvgEngine::Base
 			transform = _transform;
 		}
 
-		GameObject(int x, int y)
+		GameObject(float x, float y)
 		{
 			transform = Render::Rect();
 			transform.x = x;
